Break out recognize
