+++ src/proxy/http/profiles/router.rs
@@ -1,5 +1,3 @@
-extern crate linkerd2_router as rt;
-
 use futures::{Async, Poll, Stream};
 use http;
 use std::hash::Hash;
@@ -10,6 +8,7 @@ use dns;
 use proxy::Error;
 use svc;
 
+use super::recognize::*;
 use super::*;
 
 // A router which routes based on the `dst_overrides` of the profile or, if
@@ -96,82 +95,6 @@ where
     concrete_router: Option<ConcreteRouter<Target, Inner::Value, InnerBody>>,
     router: RouteRouter<Target, Target::Output, RouteMake::Value, RouteBody>,
     default_route: Route,
-}
-
-#[derive(Clone)]
-pub struct RouteRecognize<T> {
-    target: T,
-    routes: Vec<(RequestMatch, Route)>,
-    default_route: Route,
-}
-
-#[derive(Clone)]
-pub struct ConcreteDstRecognize<T> {
-    target: T,
-    dst_overrides: Vec<WeightedAddr>,
-    // A weighted index of the `dst_overrides` weights.  This must only be
-    // None if `dst_overrides` is empty.
-    distribution: Option<WeightedIndex<u32>>,
-}
-
-impl<Body, T> rt::Recognize<http::Request<Body>> for RouteRecognize<T>
-where
-    T: WithRoute + Clone,
-    T::Output: Clone + Eq + Hash,
-{
-    type Target = T::Output;
-
-    fn recognize(&self, req: &http::Request<Body>) -> Option<Self::Target> {
-        for (ref condition, ref route) in &self.routes {
-            if condition.is_match(&req) {
-                trace!("using configured route: {:?}", condition);
-                return Some(self.target.clone().with_route(route.clone()));
-            }
-        }
-
-        trace!("using default route");
-        Some(self.target.clone().with_route(self.default_route.clone()))
-    }
-}
-
-impl<T> ConcreteDstRecognize<T> {
-    fn new(target: T, dst_overrides: Vec<WeightedAddr>) -> Self {
-        let distribution = Self::make_dist(&dst_overrides);
-        ConcreteDstRecognize {
-            target,
-            dst_overrides,
-            distribution,
-        }
-    }
-
-    fn make_dist(dst_overrides: &Vec<WeightedAddr>) -> Option<WeightedIndex<u32>> {
-        let mut weights = dst_overrides.iter().map(|dst| dst.weight).peekable();
-        if weights.peek().is_none() {
-            // Weights list is empty.
-            None
-        } else {
-            Some(WeightedIndex::new(weights).expect("invalid weight distribution"))
-        }
-    }
-}
-
-impl<Body, T> rt::Recognize<http::Request<Body>> for ConcreteDstRecognize<T>
-where
-    T: WithAddr + Clone + Eq + Hash,
-{
-    type Target = T;
-
-    fn recognize(&self, _req: &http::Request<Body>) -> Option<Self::Target> {
-        match self.distribution {
-            Some(ref distribution) => {
-                let mut rng = rand::thread_rng();
-                let idx = distribution.sample(&mut rng);
-                let addr = self.dst_overrides[idx].addr.clone();
-                Some(self.target.clone().with_addr(addr))
-            }
-            None => Some(self.target.clone()),
-        }
-    }
 }
 
 impl<G, Inner, RouteLayer, RouteBody, InnerBody> svc::Layer<Inner>
@@ -268,11 +191,7 @@ where
         make.insert(default_route.clone(), stack.make(&default_route));
 
         let router = rt::Router::new_fixed(
-            RouteRecognize {
-                target: target.clone(),
-                routes: Vec::new(),
-                default_route: self.default_route.clone(),
-            },
+            RouteRecognize::new(target.clone(), Vec::new(), self.default_route.clone()),
             make,
         );
 
@@ -407,11 +326,7 @@ where
         }
 
         let router = rt::Router::new_fixed(
-            RouteRecognize {
-                target: self.target.clone(),
-                routes: routes.routes,
-                default_route: self.default_route.clone(),
-            },
+            RouteRecognize::new(self.target.clone(), routes.routes, self.default_route.clone()),
             make,
         );
